Define rgb color space (#1)
* WIP: Add LUV->XYZ function

It runs! That's probably about the best thing I could say about it

I verified the outputs of the LUV->XYZ manually, using an online
calculator. There's a _very_ informal set of tests that should give me peace of
mind as I clean the file up

Some things that need to be done:

- There's a good bit of duplication, not just in structure but also in
  calculation; many of these functions are called multiple times with the same
  inputs, just to return the same results

- The function APIs are kind of inconsistent, I'd like to unify them a bit

- I don't know that I'm a fan of most of the function names or the constant
  names

- *REFERENCE-WHITE/XYZ* is just one value that can be used as a reference white,
   I'll look into alternatives, and maybe define a bunch of them as named
   constants. This one in particular is called D50

* gfxmath: Use standard-objects instead of structure-objects.

* Use proper equality predicate for math object constants so new gfxmath objects are comparable.

* gfxmath: export types

* origin: Simplify mat4 inversion with a macrolet.

* origin: Fix improper mat4 invert test.

* origin: minor change.

* gfxmath: minor change.

* origin: Simply dmat4 inversion with a macrolet.

* origin: add determinant for mat3/dmat3.

* origin: Refactor mat4/dmat4 inversion.

* origin: Add mat3/dmat3 inversion.

* origin: declare speed optimization for mat3/mat4 inversion.

* origin: Export mat3/dmat3 invert!/invert.

* origin: add consing invert functions for mat3/dmat3

* Migrate LUV->XYZ from gfxmath to origin

Also make explicit the use of single floats in constant float expressions

The whole chunk of code still needs to be cleaned up, it's just that now it uses
the proper library

* Make function names more descriptive

Not all of them, mind you, though that's the overall intent

* Pare down number of function calls

Now each of the functions is called no more than once. I'm leaving in the %x and
%z functions (I know, I know) because I'm not done yet and want to keep them
around for reference. The "test" still passes, so that's a plus

* Rework how RGB color spaces are defined

This removes some of the generality of DEFINE-COLOR-SPACE in exchange for
keeping RGB-related code close together:

 - Rework DEFINE-COLOR-SPACE to be RGB-specific. It is now renamed to
   DEFINE-RGB-COLOR-SPACE to reflect that change. Now, in addition to defining a
   class, it also adds an entry to -RGB-CHROMATICITY-COORDINATES- keyed to the
   class' name (as a symbol)

 - The file color-spaces.lisp was renamed rgb-color-space.lisp, and
   rgb-metadata.lisp was partially merged into it

 - rgb-metadata.lisp was renamed to rgb-transformation-matrix.lisp, since the
   transformation matrices are the only things really left in the file

 - xyz-color-space.lisp was added to hold the definition for the class XYZ. The
   class definition for XYZ no longer uses DEFINE-COLOR-SPACE, it's a bare
   DEFCLASS form instead

It might be worth keeping DEFINE-COLOR-SPACE the way it is and creating a
wrapper macro around it for DEFINE-RGB-COLOR-SPACE. I'll look into that soon

* Restore DEFINE-COLOR-SPACE and rework previous commit in terms of it

Yeah, there was really no need to have overhauled DEFINE-COLOR-SPACE as
completely as I did, wrapping it provides a much more modular interface

* Change DEFINE-RGB-COLOR-SPACE as a MACROLET

It's used in exactly one place, we can stand to limit its scope

Co-authored-by: Catherine Tower <catherine@towernet.dev>