Finalize (de)linearation and fix chromatic adaptation.
We do not have to perform chromatic adaptation in the conversion methods, since the transformation
matrix chosen for the conversion already has that information encoded into it.

However, we must set the resultant color's standard-illuminant slot to reflect this automatic
chromatic adaptation that has occured via matrix multiplication.

As such, and along with the linearization/delinearization abstractions, the conversion routines
become incredibly concise.