Rename references to the concept of a 'Standard Illuminant' to just 'Illuminant'.
This is because it is more concise, and not all of them are standard, or published by the same
standards body.

Additionally, they are all named as keyword symbols now. This makes it more convenient to deal with
across package boundaries, since they are not objects, and just denote chromaticity coordinates of a
reference white point.