Large refactor that uses double-precision floating-point for all color calculations. Color object printing coerces elements to single-precision to visualize how the image array, which will be single-precision, will behave.
Additionally, this commit splits things into 4 different packages:

COLOR: Everything related to colors and operations on them.

IMAGE: Everything related to images (an array of pixels - this is very much a work in progress
still)

BASE: Everything common to color and image.

DEV: A Tiny package for convenience at the REPL while testing. All it does is import the above 3
packages into its namespace, and defines a CONTEXT bound to *CONTEXT* that inherits from the default
builtin context.