Constructors now take required arguments for their channels instead of keyword arguments.
This simplifies most user call sites, at the expense of nothing (the channel names are implied by
the name of the constructor).

Additionally, this adds a DEFAULT generic function, that is capable of constructing a "default"
color, for the purpose of conversion targets, or when the channel values do not matter.