Prevent the need for with-context, as suggested by Catherine Tower.
This completely refactors the context object, making it more performant.

There is now a WITH-THREADS macro, intended to be wrapped around a user's entry point, and supplied
a desirable number of threads to use for internal algorithms that are designed to be parallel.
Without WITH-THREADS, everything will run sequentiallly. This makes REPL testing a bit more concise.