Add transform-path-finder
So my idea is thus: a color conversion may happen in several steps, and the
steps are known ahead of time. Rather than writing N methods where we arrange
the conversion steps individually, why not model the conversion steps as data
and have the machine put the steps in the right order? It's a work in progress
to be sure, but I think it's promising