Weight non-XYZ successor edges in the color space graph higher than XYZ.
A higher weight means the edge will cost more when computing the shortest path. This is what we
want, because XYZ is going to be our preferred intermediate conversion step, as it has a very wide
gamut.

In the future, we may want to add a slot to the COLOR class to give different weights to different
models, if we need to bias the paths that are calculated any further.