Add a FIND-CONVERSION-PATH function that finds a path from one model to another.
This is more complicated than simply finding the shortest path, because we actually want to find the
shortest path of at least length 1. This is because, for example, if we were to look for a path from
RGB to RGB, we wouldn't want it to return NIL, but to actually go through the graph to its
successors and find a shortest path back to itself. In the case of RGB->RGB, this would mean a path
of RGB -> XYZ -> RGB, which is correct for our purposes.