Fix pooling from leaking lots of memory and effectively not working at all in a specific scenario.
The scenario where it did not work and leaked memory, was when WITH-POOL-COLOR did not specify a
space. The problem was, that we retrieved a color with a NIL space, which told the constructor to
use the default space, but then when we returned the color back to the pool when we were finished,
we queried the color for its space, and ensured a pool existed with the queried model/space
combination of that color, but the pool that it was originally a member of was the (model . nil)
pool.

The fix was quite simple. Pass a space argument to PUT-POOL-COLOR that was the same space passed to
GET-POOL-COLOR, and do not query for the color's space in GET-POOL-COLOR, instead using the passed
in space.

As an additional optimization, we also pass the model to GET-POOL-COLOR, to prevent us having to
query for that.